% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Title Odds Ratios and 95% Confidence Intervals}
\usage{
OR_95CI(coef, se, siglevel = 0.05, roundto = 2)
}
\arguments{
\item{coef}{beta coefficient}

\item{se}{standard error}

\item{siglevel}{significance level}

\item{roundto}{round to 2 decimal places}
}
\description{
Title Odds Ratios and 95% Confidence Intervals
}
